// src/lib/api/generated/model/modelProfileServiceListModelProfilesParams.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Aisphere Hub API
 * OpenAPI spec version: v1
 */

export type ModelProfileServiceListModelProfilesParams = {
pageSize?: number;
pageToken?: string;
query?: string;
/**
 * filter by status
 */
status?: string;
/**
 * filter by provider
 */
provider?: string;
};