// src/lib/api/generated/model/v1CreateModelProfileRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Aisphere Hub API
 * OpenAPI spec version: v1
 */
import type { V1CreateModelProfileRequestLabels } from './v1CreateModelProfileRequestLabels';
import type { V1ModelProfileLimits } from './v1ModelProfileLimits';

export interface V1CreateModelProfileRequest {
  allowedTools?: string[];
  apiFormat: string;
  description?: string;
  displayName?: string;
  endpoint: string;
  id: string;
  labels?: V1CreateModelProfileRequestLabels;
  limits?: V1ModelProfileLimits;
  metadata?: string;
  model?: string;
  orgId?: string;
  projectId?: string;
  provider: string;
  reasoning?: string;
  secretRef?: string;
  status?: string;
  upstreamModel: string;
  upstreamPath?: string;
  version?: string;
}